# server/crates/networkd/static/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FlightPortrait network API",
...
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://data.flightportrait.com"
    }
  ],
  "paths": {
...
    "/v1/aircraft": {
      "get": {
        "tags": [
          "Live"
        ],
        "summary": "Aircraft",
        "description": "Aircraft the network hears right now, readsb field dialect. With `bbox=minLon,minLat,maxLon,maxLat` only the aircraft with a position inside it (a box whose west edge is east of its east edge crosses the antimeridian); `total` stays the whole network. 503 if the snapshot is older than 60 seconds. Rate: 300 per 600 s (bucket `aircraft`). Cache: 10 s edge, 5 s browser.",
        "operationId": "aircraft",
        "parameters": [
          {
            "name": "bbox",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 80
                },
                {
                  "type": "null"
                }
              ],
              "description": "minLon,minLat,maxLon,maxLat; only aircraft with a position inside.",
              "title": "Bbox"
            },
            "description": "minLon,minLat,maxLon,maxLat; only aircraft with a position inside."
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "generated_at": {
                      "type": "number",
                      "description": "Snapshot time, unix seconds UTC."
                    },
                    "total": {
                      "type": "integer",
                      "description": "Aircraft in the whole snapshot, before any bbox filter."
                    },
                    "with_position": {
                      "type": "integer",
                      "description": "Of those, aircraft with a position fix."
                    },
                    "aircraft": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "hex": {
                            "type": "string",
                            "description": "ICAO 24-bit address, lowercase hex."
                          },
                          "flight": {
                            "type": "string",
                            "description": "Callsign, trailing spaces stripped. Absent when not broadcast."
                          },
                          "t": {
                            "type": "string",
                            "description": "ICAO type designator."
                          },
                          "r": {
                            "type": "string",
                            "description": "Registration."
                          },
                          "lat": {
                            "type": "number",
                            "description": "WGS84 degrees."
                          },
                          "lon": {
                            "type": "number",
                            "description": "WGS84 degrees."
                          },
                          "alt_baro": {
                            "description": "Barometric altitude, feet, or the string \"ground\".",
                            "oneOf": [
                              {
                                "type": "integer"
                              },
                              {
                                "const": "ground"
                              }
                            ]
                          },
                          "gs": {
                            "type": "number",
                            "description": "Ground speed, knots."
                          },
                          "track": {
                            "type": "number",
                            "description": "Track, degrees true."
                          },
                          "category": {
                            "type": "string",
                            "description": "ADS-B emitter category (A0-D7)."
                          },
                          "squawk": {
                            "type": "string",
                            "description": "Transponder code, 4 octal digits."
                          },
                          "seen": {
                            "type": "number",
                            "description": "Seconds since the last message."
                          },
                          "seen_pos": {
                            "type": "number",
                            "description": "Seconds since the last position."
                          },
                          "baro_rate": {
                            "type": "integer",
                            "description": "Vertical rate, feet per minute, positive climbing."
                          },
                          "emergency": {
                            "type": "string",
                            "description": "Transponder emergency state: none, general, lifeguard, minfuel, nordo, unlawful, downed, reserved."
                          }
                        },
                        "required": [
                          "hex"
                        ],
                        "description": "readsb passthrough dialect. Optional fields are absent when unknown, never null."
                      }
                    }
                  },
                  "required": [
                    "generated_at",
                    "total",
                    "with_position",
                    "aircraft"
                  ]
                },
                "example": {
                  "generated_at": 1787924061.0,
                  "total": 191,
                  "with_position": 184,
                  "aircraft": [
                    {
                      "hex": "76cd06",
                      "flight": "SIA123",
                      "t": "A359",
                      "r": "9V-SHF",
                      "lat": 1.5,
                      "lon": 103.8,
                      "alt_baro": 6325,
                      "gs": 285.9,
                      "track": 85.4,
                      "category": "A5",
                      "squawk": "2136",
                      "seen": 0.2,
                      "seen_pos": 1.1,
                      "baro_rate": 1856,
                      "emergency": "none"
                    }
                  ]
                }
              }
            }
          },
          "422": {
            "description": "Malformed input.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Machine code: not_found, not_observed, invalid_request, rate_limited, stale_snapshot, artifact_unavailable, upstream_unavailable."
                    },
                    "detail": {
                      "type": "string",
                      "description": "Human-readable explanation."
                    }
                  },
                  "required": [
                    "error",
                    "detail"
                  ]
                }
              }
            }
          },
          "429": {
            "description": "Rate limited. Wait Retry-After seconds. Headers: Retry-After, RateLimit-Limit, RateLimit-Remaining, RateLimit-Reset.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Machine code: not_found, not_observed, invalid_request, rate_limited, stale_snapshot, artifact_unavailable, upstream_unavailable."
                    },
                    "detail": {
                      "type": "string",
                      "description": "Human-readable explanation."
                    }
                  },
                  "required": [
                    "error",
                    "detail"
                  ]
                },
                "example": {
                  "error": "rate_limited",
                  "detail": "slow down"
                }
              }
            }
          },
          "503": {
            "description": "Live snapshot older than 60 seconds (stale_snapshot) or a required history artifact is not loaded (artifact_unavailable). Retry-After is set.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Machine code: not_found, not_observed, invalid_request, rate_limited, stale_snapshot, artifact_unavailable, upstream_unavailable."
                    },
                    "detail": {
                      "type": "string",
                      "description": "Human-readable explanation."
                    }
                  },
                  "required": [
                    "error",
                    "detail"
                  ]
                }
              }
            }
          }
        },
        "x-stability": "stable"
      }
    },
    "/v1/estimated": {
      "get": {
        "tags": [
          "Live"
        ],
        "summary": "Estimated positions",
        "description": "Aircraft the network stopped hearing while cruising, drawn where they most likely are: flown on from the last observed position toward the destination their callsign's route names, at the last observed speed (holding the last track 10 minutes, then turning toward the destination). Every entry is an estimate, never an observation: `estimated` is always true, `last_seen` is the last real position, `alt_baro` and `gs` are the last observed values. An aircraft appears 90 s after it was last heard and leaves when it is heard again, nears its destination, or its flight time runs out. Kept in memory, never archived or exported. Scored every night against the previous day's real coverage gaps; `accuracy` is the latest night's result for the method in use (median error by gap length), null before the first. Rate: 300 per 600 s (bucket `estimated`). Cache: 15 s edge, 10 s browser.",
        "operationId": "estimated",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "generated_at": {
                      "type": "number",
                      "description": "Unix seconds the estimates are for."
                    },
                    "method": {
                      "type": "string",
                      "description": "How positions are estimated."
                    },
                    "accuracy": {
                      "type": [
                        "object",
                        "null"
                      ],
                      "description": "The latest night's measured accuracy: day, n, median_km, by_gap_min."
                    },
                    "aircraft": {
                      "items": {
                        "type": "object"
                      },
                      "type": "array",
                      "description": "Estimated aircraft; every entry is an estimate, never an observation."
                    }
                  },
                  "type": "object",
                  "required": [
                    "generated_at",
                    "method",
                    "aircraft"
                  ]
                },
                "example": {
                  "generated_at": 1790140000.0,
                  "method": "converge-to-destination",
                  "accuracy": {
                    "day": "2026-09-22",
                    "n": 208,
                    "median_km": 2.3,
                    "by_gap_min": {
                      "5-15": {
                        "n": 153,
                        "median_km": 1.4
                      },
                      "60-120": {
                        "n": 43,
                        "median_km": 76.9
                      }
                    }
                  },
                  "aircraft": [
                    {
                      "hex": "4ca8e4",
                      "flight": "RYR1153",
                      "t": "B738",
                      "r": "9H-QDS",
                      "category": "A3",
                      "lat": 43.1021,
                      "lon": 11.9403,
                      "track": 312.4,
                      "alt_baro": 36000,
                      "gs": 452.0,
                      "estimated": true,
                      "last_seen": {
                        "at": 1790139412.0,
                        "lat": 42.2,
                        "lon": 13.1
                      },
                      "destination": "PSA",
                      "eta": 1790140700
                    }
                  ]
                }
              }
            }
          },
          "429": {
            "description": "Rate limited. Wait Retry-After seconds. Headers: Retry-After, RateLimit-Limit, RateLimit-Remaining, RateLimit-Reset.",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Machine code: not_found, not_observed, invalid_request, rate_limited, stale_snapshot, artifact_unavailable, upstream_unavailable."
                    },
                    "detail": {
                      "type": "string",
                      "description": "Human-readable explanation."
                    }
                  },
                  "type": "object",
                  "required": [
                    "error",
                    "detail"
                  ]
                },
                "example": {
                  "error": "rate_limited",
                  "detail": "slow down"
                }
              }
            }
          }
        },
        "x-stability": "map"
      }
    },
...
    "/v1/stations": {
      "get": {
        "tags": [
          "Stations"
        ],
        "summary": "Roster",
        "description": "Public feeder list. Locations rounded to about 11 km. IPs are not stored. Rate: 120 per 600 s (bucket `stations`). Cache: 30 s edge.",
        "operationId": "stations",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "stations": {
                      "items": {
                        "properties": {
                          "id": {
                            "type": "string",
                            "description": "Public id, fp-<10 hex>. Reveals no UUID bits."
                          },
                          "label": {
                            "type": [
                              "string",
                              "null"
                            ],
                            "description": "Operator-set display label."
                          },
                          "coarse_lat": {
                            "type": [
                              "number",
                              "null"
                            ],
                            "description": "Coverage midpoint, rounded ~11 km."
                          },
                          "coarse_lon": {
                            "type": [
                              "number",
                              "null"
                            ]
                          },
                          "first_seen": {
                            "type": "string",
                            "description": "ISO 8601 UTC."
                          },
                          "last_seen": {
                            "type": "string",
                            "description": "ISO 8601 UTC."
                          },
                          "online": {
                            "type": "boolean"
                          }
                        },
                        "type": "object",
                        "required": [
                          "id",
                          "first_seen",
                          "last_seen",
                          "online"
                        ]
                      },
                      "type": "array"
                    }
                  },
                  "type": "object",
                  "required": [
                    "stations"
                  ]
                },
                "example": {
                  "stations": [
                    {
                      "id": "fp-a1b2c3d4e5",
                      "coarse_lat": 1.4,
                      "coarse_lon": 103.8,
                      "first_seen": "2026-08-01T00:00:00+00:00",
                      "last_seen": "2026-08-29T12:00:00+00:00",
                      "online": true
                    }
                  ]
                }
              }
            }
          },
          "429": {
            "description": "Rate limited. Wait Retry-After seconds. Headers: Retry-After, RateLimit-Limit, RateLimit-Remaining, RateLimit-Reset.",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Machine code: not_found, not_observed, invalid_request, rate_limited, stale_snapshot, artifact_unavailable, upstream_unavailable."
                    },
                    "detail": {
                      "type": "string",
                      "description": "Human-readable explanation."
                    }
                  },
                  "type": "object",
                  "required": [
                    "error",
                    "detail"
                  ]
                },
                "example": {
                  "error": "rate_limited",
                  "detail": "slow down"
                }
              }
            }
          }
        },
        "x-stability": "stable"
      }
    },
...
  },
...
}